requirements.txt
python-pptx==0.6.21
Pillow==10.1.0
firebase-admin==6.3.0 